still bugs in the collision groups -> enemies killing each other

+++ rust_survivors/src/enemy/components.rs
@@ -6,7 +6,7 @@ pub struct Enemy;
 
 #[derive(Component)]
 pub struct ExpGem {
-    pub value: f32,
+    pub value: i32,
 }
 
 #[derive(Component)]
